mobile/pyproject.toml
[project]
name = "mobile"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.35.0",
    "colorlog>=6.8.0",
    "pycryptodome>=3.23.0",
    "requests>=2.32.5",
    "tqdm>=4.66.0",
]
